requirements.txt
click~=8.1.7
python-dotenv~=1.0.1
PyGithub~=2.3.0
PyYAML~=6.0.1